+++ src/fft.rs
@@ -167,32 +167,32 @@ is desirable for better locality of memory accesses).
 pub mod radix2 {
     use enums;
 
-    pub fn forward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
+    pub fn forward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_forward(data.as_mut_ptr(), stride, n)
         })
     }
 
-    pub fn transform(data: &mut [f64], stride: u64, n: u64, sign: ::FftDirection) -> enums::Value {
+    pub fn transform(data: &mut [f64], stride: usize, n: usize, sign: ::FftDirection) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_transform(data.as_mut_ptr(), stride, n, sign.into())
         })
     }
 
-    pub fn backward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
+    pub fn backward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_backward(data.as_mut_ptr(), stride, n)
         })
     }
 
-    pub fn inverse(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
+    pub fn inverse(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_inverse(data.as_mut_ptr(), stride, n)
         })
     }
 
     /// This is decimation-in-frequency version of the radix-2 FFT function.
-    pub fn dif_forward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
+    pub fn dif_forward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_dif_forward(data.as_mut_ptr(), stride, n)
         })
@@ -201,8 +201,8 @@ pub mod radix2 {
     /// This is decimation-in-frequency version of the radix-2 FFT function.
     pub fn dif_transform(
         data: &mut [f64],
-        stride: u64,
-        n: u64,
+        stride: usize,
+        n: usize,
         sign: ::FftDirection,
     ) -> enums::Value {
         enums::Value::from(unsafe {
@@ -211,14 +211,14 @@ pub mod radix2 {
     }
 
     /// This is decimation-in-frequency version of the radix-2 FFT function.
-    pub fn dif_backward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
+    pub fn dif_backward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_dif_backward(data.as_mut_ptr(), stride, n)
         })
     }
 
     /// This is decimation-in-frequency version of the radix-2 FFT function.
-    pub fn dif_inverse(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
+    pub fn dif_inverse(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_dif_inverse(data.as_mut_ptr(), stride, n)
         })
@@ -247,8 +247,8 @@ pub mod mixed_radix {
 
     pub fn forward(
         data: &mut [f64],
-        stride: u64,
-        n: u64,
+        stride: usize,
+        n: usize,
         wavetable: &FftComplexWaveTable,
         work: &mut FftComplexWorkspace,
     ) -> enums::Value {
@@ -265,8 +265,8 @@ pub mod mixed_radix {
 
     pub fn transform(
         data: &mut [f64],
-        stride: u64,
-        n: u64,
+        stride: usize,
+        n: usize,
         wavetable: &FftComplexWaveTable,
         work: &mut FftComplexWorkspace,
         sign: ::FftDirection,
@@ -285,8 +285,8 @@ pub mod mixed_radix {
 
     pub fn backward(
         data: &mut [f64],
-        stride: u64,
-        n: u64,
+        stride: usize,
+        n: usize,
         wavetable: &FftComplexWaveTable,
         work: &mut FftComplexWorkspace,
     ) -> enums::Value {
@@ -303,8 +303,8 @@ pub mod mixed_radix {
 
     pub fn inverse(
         data: &mut [f64],
-        stride: u64,
-        n: u64,
+        stride: usize,
+        n: usize,
         wavetable: &FftComplexWaveTable,
         work: &mut FftComplexWorkspace,
     ) -> enums::Value {
@@ -352,7 +352,7 @@ pub mod real_radix2 {
     /// complex[n-1].imag  =   -data[n-1]
     /// Note that the output data can be converted into the full complex sequence using the function gsl_fft_halfcomplex_radix2_unpack described
     /// below.
-    pub fn transform(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
+    pub fn transform(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_real_radix2_transform(data.as_mut_ptr(), stride, n)
         })
@@ -360,7 +360,7 @@ pub mod real_radix2 {
 
     /// This function computes the inverse or backwards in-place radix-2 FFT of length n and stride stride on the half-complex sequence data
     /// stored according the output scheme used by gsl_fft_real_radix2. The result is a real array stored in natural order.
-    pub fn inverse(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
+    pub fn inverse(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_halfcomplex_radix2_inverse(data.as_mut_ptr(), stride, n)
         })
@@ -368,7 +368,7 @@ pub mod real_radix2 {
 
     /// This function computes the inverse or backwards in-place radix-2 FFT of length n and stride stride on the half-complex sequence data
     /// stored according the output scheme used by gsl_fft_real_radix2. The result is a real array stored in natural order.
-    pub fn backward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
+    pub fn backward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_halfcomplex_radix2_backward(data.as_mut_ptr(), stride, n)
         })
@@ -407,8 +407,8 @@ pub mod real_radix2 {
     pub fn unpack(
         halfcomplex_coefficient: &mut [f64],
         complex_coefficient: &mut [f64],
-        stride: u64,
-        n: u64,
+        stride: usize,
+        n: usize,
     ) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_halfcomplex_radix2_unpack(
