+++ src/types/interpolation.rs
@@ -67,7 +67,7 @@ impl InterpAccel {
     /// This function performs a lookup action on the data array x_array of size size, using the
     /// given accelerator a. This is how lookups are performed during evaluation of an
     /// interpolation. The function returns an index i such that `x_array[i] <= x < x_array[i+1]`.
-    pub fn find(&mut self, x_array: &[f64], x: f64) -> u64 {
+    pub fn find(&mut self, x_array: &[f64], x: f64) -> usize {
         unsafe { sys::gsl_interp_accel_find(&mut self.0, x_array.as_ptr(), x_array.len() as _, x) }
     }
 }
@@ -86,7 +86,7 @@ impl Interp {
     /// let interp_type = InterpType::linear();
     /// let interp = Interp::new(&interp_type, 2).expect("Failed to initialize `Interp`...");
     /// ```
-    pub fn new(t: &InterpType, size: u64) -> Option<Interp> {
+    pub fn new(t: &InterpType, size: usize) -> Option<Interp> {
         let tmp = unsafe { sys::gsl_interp_alloc(t.t, size) };
 
         if tmp.is_null() {
@@ -246,7 +246,7 @@ pub struct Spline {
 }
 
 impl Spline {
-    pub fn new(t: &InterpType, size: u64) -> Option<Spline> {
+    pub fn new(t: &InterpType, size: usize) -> Option<Spline> {
         let tmp = unsafe { sys::gsl_spline_alloc(t.t, size) };
 
         if tmp.is_null() {
